+++ dbaccess/gormdao/cond.go
@@ -1,4 +1,4 @@
-package genericdao
+package gormdao
 
 import (
 	"fmt"
