+++ dbaccess/gormdao/cond_test.go
@@ -1,4 +1,4 @@
-package genericdao
+package gormdao
 
 import (
 	"testing"
